fix: rp_ipa service, new diccionary ad function to check words with 'æ' | delete: unused files phonetic*
