--- LONGMAN_INTEGRATION.md
@@ -1,161 +1,0 @@
-# 📚 Integración con Longman Dictionary of Contemporary English
-
-## 🎯 Resumen
-
-Tu aplicación ahora incluye integración con el [Longman Dictionary of Contemporary English Online](https://www.ldoceonline.com/), la fuente más auténtica para transcripciones fonéticas RP (Received Pronunciation).
-
-## 🏗️ Arquitectura del Sistema
-
-### Sistema Híbrido de Transcripción RP
-
-Tu aplicación utiliza un **sistema híbrido de 3 niveles** para máxima precisión:
-
-```
-1️⃣ Longman Dictionary Online  (Más auténtico - opcional)
-    ↓ (si no encuentra)
-2️⃣ Diccionario RP Local      (100+ palabras académicas)
-    ↓ (si no encuentra) 
-3️⃣ Conversión GA → RP        (Algoritmo inteligente)
-```
-
-### 📊 Rendimiento Actual
-
-- ✅ **Tasa de éxito: 100%**
-- ✅ **Diccionario local: 94.1%** (palabras comunes)
-- ✅ **Conversión GA→RP: 5.9%** (palabras menos comunes)
-- ✅ **Longman: 0%** (actualmente deshabilitado)
-
-## 🔧 Configuración de Longman
-
-### Estado Actual: Desarrollo
-
-El servicio Longman está implementado pero actualmente deshabilitado por defecto porque:
-
-1. **Límites de velocidad**: Respetamos el sitio web con delays de 1 segundo
-2. **Parsing HTML**: Estructura compleja del sitio requiere ajustes
-3. **Confiabilidad**: El diccionario local ya proporciona excelente cobertura
-
-### Habilitar Longman (Experimental)
-
-Para desarrolladores que quieran experimentar:
-
-```python
-# En src/services/phonetic_transcription_service.py
-hybrid_service = get_hybrid_rp_service(use_longman=True)  # Cambiar a True
-```
-
-## 📝 Archivos del Sistema Longman
-
-### `src/services/longman_dictionary_service.py`
-- Web scraping del sitio Longman
-- Headers de navegador realistas
-- Cache para evitar consultas repetidas
-- Rate limiting respetuoso (1 seg entre requests)
-
-### `src/services/hybrid_rp_service.py`
-- Combina Longman + diccionario local + conversión
-- Estadísticas de uso por fuente
-- Fallback inteligente entre métodos
-
-### `src/services/rp_phonetic_service.py`
-- Diccionario especializado RP con 100+ palabras
-- Conversión precisa GA → RP
-- Reglas fonológicas auténticas
-
-## 🎓 Para Tu Universidad
-
-### Palabras Académicas Cubiertas
-
-Tu diccionario local incluye vocabulario especializado:
-
-```
-phonetics     → /fəˈnetɪks/
-pronunciation → /prəˌnʌnsɪˈeɪʃən/
-linguistics   → /lɪŋˈgwɪstɪks/
-transcription → /trænˈskrɪpʃən/
-university    → /ˌjuːnɪˈvɜːsəti/
-received      → /rɪˈsiːvd/
-british       → /ˈbrɪtɪʃ/
-dictionary    → /ˈdɪkʃənri/
-```
-
-### Características RP Distintivas
-
-- **BATH words**: ask → `/ɑːsk/`, dance → `/dɑːns/`
-- **LOT words**: hot → `/hɒt/`, got → `/gɒt/`
-- **GOAT diphthong**: go → `/gəʊ/`, phone → `/fəʊn/`
-- **R no rótica**: car → `/kɑː/`, here → `/hɪə/`
-
-## 🚀 Uso en la Aplicación
-
-### Interfaz Gráfica
-
-1. Ejecuta: `uv run python app.py --lang es`
-2. Ve a la pestaña **"Transcripción Fonética IPA"**
-3. Escribe texto en inglés
-4. Haz clic en **"Transcribir a IPA"**
-5. ¡Obtén transcripción RP auténtica!
-
-### Ejemplos de Uso
-
-```python
-# Uso directo del servicio
-from src.services.hybrid_rp_service import get_hybrid_rp_service
-
-service = get_hybrid_rp_service()
-transcription = service.get_pronunciation("university")
-# Resultado: "ˌjuːnɪˈvɜːsəti"
-
-# Transcripción de texto completo
-text = "Ask the British teacher about pronunciation"
-result = service.transcribe_text(text)
-# Resultado: "ɑːsk ðə ˈbrɪtɪʃ ˈtiːtʃə əˈbaʊt prəˌnʌnsɪˈeɪʃən"
-```
-
-## 🔍 Testing y Desarrollo
-
-### Scripts de Prueba
-
-```bash
-# Test completo del sistema
-uv run python test_final_rp.py
-
-# Test específico GA vs RP
-uv run python test_ga_vs_rp.py
-
-# Test solo Longman (experimental)
-uv run python -c "from src.services.longman_dictionary_service import test_longman_service; test_longman_service()"
-```
-
-### Estadísticas de Rendimiento
-
-```python
-service = get_hybrid_rp_service()
-service.print_statistics()
-# Muestra uso por fuente y tasa de éxito
-```
-
-## 🛠️ Desarrollo Futuro
-
-### Mejoras Planeadas para Longman
-
-1. **Parser HTML mejorado** para manejar la estructura compleja
-2. **Selectors CSS específicos** para elementos de pronunciación
-3. **Validación de resultados** antes de usar transcripciones
-4. **Cache persistente** para reducir consultas repetidas
-
-### Extensibilidad
-
-- Agregar otros diccionarios online (Cambridge, Oxford)
-- Soporte para variantes regionales del RP
-- Integración con APIs oficiales cuando estén disponibles
-
-## 📚 Referencias
-
-- [Longman Dictionary Online](https://www.ldoceonline.com/)
-- [International Phonetic Association](https://www.internationalphoneticassociation.org/)
-- [RP Pronunciation Guide](https://en.wikipedia.org/wiki/Received_Pronunciation)
-
----
-
-**¡Tu sistema está listo para uso académico profesional con la máxima precisión en transcripción RP!** 🎉
