--- requirements.txt
@@ -1,5 +1,0 @@
-googletrans==4.0.0rc1
-pdfplumber==0.11.7
-python-docx==1.2.0
-fpdf2==2.8.3
-spacy==3.8.7
